Code Done!

+++ Project1Library/Data/RockPaperScissor.cs
@@ -14,7 +14,7 @@ namespace Project1Library.Data
         public int Vinst { get; set; }
         public int Oavgjort { get; set; }
         public int Förlust { get; set; }
-        public double Genomsnitt { get; set; }
+        public decimal Genomsnitt { get; set; }
         public DateTime Datum { get; set; }
         public string SpelarensDrag { get; set; }
         public string DatornsDrag { get; set; }
